requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-datarobot>=0.8.13,<0.9

datarobot==3.4.0
datarobot-predict>=1.9.6,<1.10
datarobotx-idp>=0.2.13,<0.3

python-dotenv==1.0.1
pydantic>=2.9.2,<2.10
pydantic-settings>=2.5.2,<2.6
eval-type-backport>=0.2.0,<0.3

jupyterlab>=4.2.5,<5
papermill>=2.6.0,<3

babel>=2.16,<3
pandas>=2.2.3,<3

pytest==8.0.1
pytest-cov==4.1.0
pytest-mock==3.12.0
mypy==1.11.2
types-PyYAML==6.0.12.20240917
ruff==0.6.9
openai
